Rename to_syrup and from_syrup to better reflect behavour and add to_syrup
The to_syrup and from_syrup didn't work on encoded syrup data, but
worked on the syrup.Record data type. This was a bit confusing in the
code so were renamed.

Due to the rename, this also let us add `to_syrup` which takes the syrup
record and actually encodes it to syrup encoded octets.