Remove op:bootstrap and fix op:abort tests
This removes the op:bootstrap operation in favor of always exporting the
bootstrap object at position 0. This change was agreed on in the
November 2023 OCapN pre-standardization meeting.

This also fixes a bug with the op:abort tests which were actually not
working as intended. They neglected to actually emit the op:abort that
were created.