Change URI encoding in tcp-testing-only netlayer
The port and hostname were placed in the URI, however their
placement wasn't how ports are usually encoding, creating an
invalid URI. This instead uses a unauthenticated random string
as the designator, and the port and host are moved to the hints
section which is now supported in OCapN nodes.

Prior to this hints were underspecified in the spec, this is still
the case, the test suite for now assumes an OCapN struct, this is
represented using the syrup hash type.